@@ -1,13 +1,14 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import ActionSheet from './ActionSheet';
 
 export default class ActionSheetProvider extends React.Component {
   static propTypes = {
-    children: React.PropTypes.any,
+    children: PropTypes.any,
   };
 
   static childContextTypes = {
-    showActionSheetWithOptions: React.PropTypes.func,
+    showActionSheetWithOptions: PropTypes.func,
   };
 
   getChildContext() {
@@ -19,7 +20,7 @@ export default class ActionSheetProvider extends React.Component {
 
   render() {
     return (
-      <ActionSheet ref={component => this._actionSheetRef = component}>
+      <ActionSheet ref={component => (this._actionSheetRef = component)}>
         {React.Children.only(this.props.children)}
       </ActionSheet>
     );
