@@ -5,11 +5,7 @@ import { ActionSheetIOS, View } from 'react-native';
 
 export default class ActionSheet extends React.Component {
   render() {
-    return (
-      <View style={{ flex: 1 }}>
-        {React.Children.only(this.props.children)}
-      </View>
-    );
+    return <View style={{ flex: 1 }}>{React.Children.only(this.props.children)}</View>;
   }
 
   showActionSheetWithOptions(options, onSelect) {
