@@ -2,12 +2,9 @@ import Expo from 'expo';
 import React from 'react';
 import { StyleSheet, Text, TouchableOpacity, View } from 'react-native';
 import { Entypo } from '@expo/vector-icons';
-import {
-  ActionSheetProvider,
-  connectActionSheet,
-} from '@expo/react-native-action-sheet';
+import { ActionSheetProvider, connectActionSheet } from '@expo/react-native-action-sheet';
 
-class AppContainer extends React.Component {
+export default class AppContainer extends React.Component {
   render() {
     return (
       <ActionSheetProvider>
@@ -17,21 +14,18 @@ class AppContainer extends React.Component {
   }
 }
 
-@connectActionSheet class App extends React.Component {
+@connectActionSheet
+class App extends React.Component {
   render() {
     return (
       <View style={styles.container}>
         <View style={{ marginBottom: 30 }}>
           <Text style={{ textAlign: 'center' }}>
-            Hello! This is a simple example app to demonstrate
-            @exponent/react-native-action-sheet.
+            Hello! This is a simple example app to demonstrate @expo/react-native-action-sheet.
           </Text>
         </View>
 
-        <Entypo.Button
-          name="code"
-          backgroundColor="#3e3e3e"
-          onPress={this._onOpenActionSheet}>
+        <Entypo.Button name="code" backgroundColor="#3e3e3e" onPress={this._onOpenActionSheet}>
           <Text style={{ fontSize: 15, color: '#fff' }}>Open action sheet</Text>
         </Entypo.Button>
       </View>
@@ -64,5 +58,3 @@ const styles = StyleSheet.create({
     justifyContent: 'center',
   },
 });
-
-Expo.registerRootComponent(AppContainer);
